bundle/direct/dstate: fold DMS state read into StateDB.Open
Address review: drop the StateReader interface and its readLocalDatabase
duplicate of StateDB.Open. Open now takes a deployment-metadata-service
client (nil = file-only) and overlays DMS resource state on top of the
local identity (lineage/serial) when DMS owns the deployment. The
version-completion gate and resource fetch move into the dstate package.
