+++ cmd/bundle/generate/dashboard.go
@@ -394,7 +394,7 @@ func (d *dashboard) runForResource(ctx context.Context, b *bundle.Bundle) {
 	var state statemgmt.ExportedResourcesMap
 	if stateDesc.Engine.IsDirect() {
 		_, localPath := b.StateFilenameDirect(ctx)
-		if err := b.DeploymentBundle.StateDB.Open(ctx, localPath, dstate.WithRecovery(true), dstate.WithWrite(false)); err != nil {
+		if err := b.DeploymentBundle.StateDB.Open(ctx, localPath, dstate.WithRecovery(true), dstate.WithWrite(false), nil); err != nil {
 			logdiag.LogError(ctx, err)
 			return
 		}
